// types/supabase.types.ts
export type Json =
    | string
    | number
    | boolean
    | null
    | { [key: string]: Json | undefined }
    | Json[]

export type Database = {
// ...
    public: {
        Tables: {
// ...
            orders: {
                Row: {
                    created_at: string | null
                    id: string
                    is_partial_order: boolean
                    items_json: Json | null
                    parent_order_id: string | null
                    status: string
                    subtotal: number
                    table_id: string | null
                    tax: number
                    tax_percentage: number
                    tip: number
                    tip_percentage: number
                    total: number
                    total_discounts: number | null
                    updated_at: string | null
                    waiter_id: string | null
                }
                Insert: {
                    created_at?: string | null
                    id?: string
                    is_partial_order?: boolean
                    items_json?: Json | null
                    parent_order_id?: string | null
                    status?: string
                    subtotal: number
                    table_id?: string | null
                    tax: number
                    tax_percentage: number
                    tip?: number
                    tip_percentage?: number
                    total: number
                    total_discounts?: number | null
                    updated_at?: string | null
                    waiter_id?: string | null
                }
                Update: {
                    created_at?: string | null
                    id?: string
                    is_partial_order?: boolean
                    items_json?: Json | null
                    parent_order_id?: string | null
                    status?: string
                    subtotal?: number
                    table_id?: string | null
                    tax?: number
                    tax_percentage?: number
                    tip?: number
                    tip_percentage?: number
                    total?: number
                    total_discounts?: number | null
                    updated_at?: string | null
                    waiter_id?: string | null
                }
                Relationships: [
                    {
                        foreignKeyName: "orders_parent_order_id_fkey"
                        columns: ["parent_order_id"]
                        isOneToOne: false
                        referencedRelation: "orders"
                        referencedColumns: ["id"]
                    },
                    {
                        foreignKeyName: "orders_table_id_fkey"
                        columns: ["table_id"]
                        isOneToOne: false
                        referencedRelation: "tables"
                        referencedColumns: ["id"]
                    },
                    {
                        foreignKeyName: "orders_waiter_id_fkey"
                        columns: ["waiter_id"]
                        isOneToOne: false
                        referencedRelation: "profiles"
                        referencedColumns: ["id"]
                    },
                ]
            }
// ...
            tables: {
                Row: {
                    created_at: string | null
                    id: string
                    number: number
                    status: string
                    updated_at: string | null
                    waiter_id: string | null
                }
                Insert: {
                    created_at?: string | null
                    id?: string
                    number: number
                    status?: string
                    updated_at?: string | null
                    waiter_id?: string | null
                }
                Update: {
                    created_at?: string | null
                    id?: string
                    number?: number
                    status?: string
                    updated_at?: string | null
                    waiter_id?: string | null
                }
                Relationships: [
                    {
                        foreignKeyName: "tables_waiter_id_fkey"
                        columns: ["waiter_id"]
                        isOneToOne: false
                        referencedRelation: "profiles"
                        referencedColumns: ["id"]
                    },
                ]
            }
        }
        Views: {
            [_ in never]: never
        }
        Functions: {
            [_ in never]: never
        }
        Enums: {
            [_ in never]: never
        }
        CompositeTypes: {
            [_ in never]: never
        }
    }
}